refactor Repository
