+++ MMEX/ContentView.swift
@@ -197,7 +197,7 @@ struct ContentView: View {
                 print("Unable to access file at URL: \(url)")
             }
             let repository = dataManager.repository
-            repository.create(sampleData: isSampleDocument)
+            repository?.create(sampleData: isSampleDocument)
             DispatchQueue.main.asyncAfter(deadline: .now() + 0.1) {
                 selectedTab = 0
             }
