+++ MMEX/DatabaseManager.swift
@@ -10,7 +10,7 @@ import SQLite
 
 class DataManager: ObservableObject {
     @Published var isDatabaseConnected = false
-    private(set) var db: Connection?
+    private(set) var repository: Repository?
     private(set) var databaseURL: URL?
 
     init() {
@@ -19,23 +19,8 @@ class DataManager: ObservableObject {
     
     func openDatabase(at databaseURL: URL) {
         self.databaseURL = databaseURL
-        if databaseURL.startAccessingSecurityScopedResource() {
-            defer { databaseURL.stopAccessingSecurityScopedResource() }
-            do {
-                db = try Connection(databaseURL.path)
-                print("Connected to database: \(databaseURL.path)")
-                setJournalModeDelete(connection: db!)
-                isDatabaseConnected = true
-            } catch {
-                print("Failed to connect to database: \(error)")
-                db = nil
-                isDatabaseConnected = false
-            }
-        } else {
-            db = nil
-            isDatabaseConnected = false
-            print("Failed to access security-scoped resource: \(databaseURL.path)")
-        }
+        repository = Repository(url: databaseURL)
+        isDatabaseConnected = repository != nil
     }
 
     func setJournalModeDelete(connection: Connection) {
@@ -72,44 +57,38 @@ class DataManager: ObservableObject {
         }
     }
     
-    func getSchemaVersion() -> Int32? {
-        if let db {
-            return db.userVersion
-        }
-        return nil
-    }
+    func getSchemaVersion() -> Int32? { repository?.userVersion }
 
     /// Closes the current database connection and resets related states.
     func closeDatabase() {
         // Nullify the connection and reset the state
-        db = nil
+        repository = nil
         isDatabaseConnected = false
         databaseURL = nil
         print("Database connection closed.")
     }
 
-    var repository                 : Repository                 { Repository(db: db) }
-    var infotableRepository        : InfotableRepository        { InfotableRepository(db: db) }
-    var currencyRepository         : CurrencyRepository         { CurrencyRepository(db: db) }
-    var currencyHistoryRepository  : CurrencyRepository         { CurrencyRepository(db: db) }
-    var accountRepository          : AccountRepository          { AccountRepository(db: db) }
-    var assetRepository            : AssetRepository            { AssetRepository(db: db) }
-    var stockRepository            : StockRepository            { StockRepository(db: db) }
-    var stockHistoryRepository     : StockRepository            { StockRepository(db: db) }
-    var categoryRepository         : CategoryRepository         { CategoryRepository(db: db) }
-    var payeeRepository            : PayeeRepository            { PayeeRepository(db: db) }
-    var transactionRepository      : TransactionRepository      { TransactionRepository(db: db) }
-    var transactionSplitRepository : TransactionSplitRepository { TransactionSplitRepository(db: db) }
-    var transactionLinkRepository  : TransactionLinkRepository  { TransactionLinkRepository(db: db) }
-    var transactionShareRepository : TransactionShareRepository { TransactionShareRepository(db: db) }
-    var scheduledRepository        : ScheduledRepository        { ScheduledRepository(db: db) }
-    var scheduledSplitRepository   : ScheduledSplitRepository   { ScheduledSplitRepository(db: db) }
-    var tagRepository              : TagRepository              { TagRepository(db: db) }
-    var tagLinkRepository          : TagLinkRepository          { TagLinkRepository(db: db) }
-    var fieldRepository            : FieldRepository            { FieldRepository(db: db) }
-    var fieldContentRepository     : FieldContentRepository     { FieldContentRepository(db: db) }
-    var attachmentRepository       : AttachmentRepository       { AttachmentRepository(db: db) }
-    var budgetYearRepository       : BudgetYearRepository       { BudgetYearRepository(db: db) }
-    var budgetTableRepository      : BudgetTableRepository      { BudgetTableRepository(db: db) }
-    var reportRepository           : ReportRepository           { ReportRepository(db: db) }
+    var infotableRepository        : InfotableRepository?        { repository?.infotableRepository }
+    var currencyRepository         : CurrencyRepository?         { repository?.currencyRepository }
+    var currencyHistoryRepository  : CurrencyRepository?         { repository?.currencyRepository }
+    var accountRepository          : AccountRepository?          { repository?.accountRepository }
+    var assetRepository            : AssetRepository?            { repository?.assetRepository }
+    var stockRepository            : StockRepository?            { repository?.stockRepository }
+    var stockHistoryRepository     : StockRepository?            { repository?.stockRepository }
+    var categoryRepository         : CategoryRepository?         { repository?.categoryRepository }
+    var payeeRepository            : PayeeRepository?            { repository?.payeeRepository }
+    var transactionRepository      : TransactionRepository?      { repository?.transactionRepository }
+    var transactionSplitRepository : TransactionSplitRepository? { repository?.transactionSplitRepository }
+    var transactionLinkRepository  : TransactionLinkRepository?  { repository?.transactionLinkRepository }
+    var transactionShareRepository : TransactionShareRepository? { repository?.transactionShareRepository }
+    var scheduledRepository        : ScheduledRepository?        { repository?.scheduledRepository }
+    var scheduledSplitRepository   : ScheduledSplitRepository?   { repository?.scheduledSplitRepository }
+    var tagRepository              : TagRepository?              { repository?.tagRepository }
+    var tagLinkRepository          : TagLinkRepository?          { repository?.tagLinkRepository }
+    var fieldRepository            : FieldRepository?            { repository?.fieldRepository }
+    var fieldContentRepository     : FieldContentRepository?     { repository?.fieldContentRepository }
+    var attachmentRepository       : AttachmentRepository?       { repository?.attachmentRepository }
+    var budgetYearRepository       : BudgetYearRepository?       { repository?.budgetYearRepository }
+    var budgetTableRepository      : BudgetTableRepository?      { repository?.budgetTableRepository }
+    var reportRepository           : ReportRepository?           { repository?.reportRepository }
 }
