+++ MMEX/Repositories/CategoryRepository.swift
@@ -10,8 +10,8 @@ import SQLite
 class CategoryRepository: RepositoryProtocol {
     typealias RepositoryData = CategoryData
 
-    let db: Connection?
-    init(db: Connection?) {
+    let db: Connection
+    init(db: Connection) {
         self.db = db
     }
