+++ MMEX/Repositories/ScheduledRepository.swift
@@ -11,8 +11,8 @@ import SQLite
 class ScheduledRepository: RepositoryProtocol {
     typealias RepositoryData = ScheduledData
 
-    let db: Connection?
-    init(db: Connection?) {
+    let db: Connection
+    init(db: Connection) {
         self.db = db
     }
