+++ MMEX/Views/Transactions/TransactionAddView2.swift
@@ -52,7 +52,7 @@ struct TransactionAddView2: View {
             
             // database level setting
             let repository = dataManager.infotableRepository
-            if let storedDefaultAccount = repository.getValue(for: InfoKey.defaultAccountID.id, as: Int64.self) {
+            if let storedDefaultAccount = repository?.getValue(for: InfoKey.defaultAccountID.id, as: Int64.self) {
                 newTxn.accountId = storedDefaultAccount
             }
         }
@@ -64,7 +64,7 @@ struct TransactionAddView2: View {
 
     func addTransaction(txn: inout TransactionData) {
         let repository = dataManager.transactionRepository
-        if repository.insert(&txn) {
+        if repository?.insert(&txn) == true {
             // id is ready after repo call
         } else {
             // TODO
@@ -76,7 +76,7 @@ struct TransactionAddView2: View {
 
         // Fetch accounts using repository and update the view
         DispatchQueue.global(qos: .background).async {
-            let loadedPayees = repository.load()
+            let loadedPayees = repository?.load() ?? []
             
             // Update UI on the main thread
             DispatchQueue.main.async {
@@ -89,7 +89,7 @@ struct TransactionAddView2: View {
         let repository = dataManager.categoryRepository
 
         DispatchQueue.global(qos: .background).async {
-            let loadedCategories = repository.load()
+            let loadedCategories = repository?.load() ?? []
             
             DispatchQueue.main.async {
                 self.categories = loadedCategories
@@ -101,7 +101,7 @@ struct TransactionAddView2: View {
         let repository = dataManager.accountRepository
 
         DispatchQueue.global(qos: .background).async {
-            let loadedAccounts = repository.loadWithCurrency()
+            let loadedAccounts = repository?.loadWithCurrency() ?? []
             
             DispatchQueue.main.async {
                 self.accounts = loadedAccounts
